+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- qemu: add compile.sh from ostools;
 - ai: paddle: first commit;
 - transformers: First commit;
 - cmake: add version.sh, samples directory;